Remove GridElement class
